% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConditionalDensityEvaluator.R
\docType{class}
\name{ConditionalDensityEvaluator}
\alias{ConditionalDensityEvaluator}
\title{ConditionalDensityEvaluator}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ConditionalDensityEvaluator
}
\description{
ConditionalDensityEvaluator
}
\section{Methods}{

\describe{
  \item{\code{initialize() }}{
    <<description>>
  }
  \item{\code{evaluate(osl, T_iter, B_iter) }}{
    Set \code{T_iter} to be a large integer (>= 1), an example value for
    this would be 100. This is the number of conditional densities we
    evaluate. Then set \code{B_iter} to an even larger number (>= 1), e.g.
    10e4. This is the number of samples drawn from our distributions.
  }
}
}

\keyword{datasets}
