% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstrainedGlm.R
\name{ConstrainedGlm.fit_new_glm}
\alias{ConstrainedGlm.fit_new_glm}
\title{Fit a new GLM
In this function we create a new instance a (constrained)
glm fit.}
\usage{
ConstrainedGlm.fit_new_glm(formula, family, data, fall_back_to_glm, ...)
}
\arguments{
\item{formula}{the formula to use for the regression}

\item{family}{the family used for fitting the GLM (binomial, etc)}

\item{data}{the data to train the glm on}

\item{fall_back_to_glm}{boolean should we fall back to traditional glm}

\item{...}{the other arguments passed to GLM}
}
\value{
a fitted glm
}
\description{
Fit a new GLM
In this function we create a new instance a (constrained)
glm fit.
}
