% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstrainedGlm.R
\name{ConstrainedGlm.predict}
\alias{ConstrainedGlm.predict}
\title{Predict using a constrained glm
In this function we predict usng an instance of a (constrained)
glm fit.}
\usage{
ConstrainedGlm.predict(constrained_glm, newdata)
}
\arguments{
\item{constrained_glm}{a fitted constrained glm instance}

\item{newdata}{the data to perform the predictions with}
}
\value{
a prediction
}
\description{
Predict using a constrained glm
In this function we predict usng an instance of a (constrained)
glm fit.
}
