% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstrainedGlm.R
\name{ConstrainedGlm.update_glm}
\alias{ConstrainedGlm.update_glm}
\title{Update Constrained logistic regression
In this function we update a previously trained instance of a (constrained)
glm fit.}
\usage{
ConstrainedGlm.update_glm(previous_glm, data, ...)
}
\arguments{
\item{previous_glm}{glm object. A previously trained GLM instance, so it can be updated}

\item{data}{the newdata to update the glm on}

\item{...}{the other arguments passed to GLM}
}
\value{
a fitted, updated glm
}
\description{
Update Constrained logistic regression
In this function we update a previously trained instance of a (constrained)
glm fit.
}
