% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossValidationRiskCalculator.R
\docType{class}
\name{CrossValidationRiskCalculator}
\alias{CrossValidationRiskCalculator}
\title{CrossValidationRiskCalculator
Class that contains various methods for calculating the crossvalidated risk
of an estimator.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CrossValidationRiskCalculator
}
\description{
CrossValidationRiskCalculator
Class that contains various methods for calculating the crossvalidated risk
of an estimator.
}
\keyword{datasets}
