% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.Stream.Simulator.R
\docType{class}
\name{Data.Stream.Simulator}
\alias{Data.Stream.Simulator}
\title{Data.Stream.Simulator
Class that serves as a very basic simulator to fill a kafka stream with data.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Data.Stream.Simulator
}
\description{
Data.Stream.Simulator
Class that serves as a very basic simulator to fill a kafka stream with data.
}
\section{Methods}{

\describe{
  \item{\code{new(kafka.host = 'imac.evionix.org:9092', kafka.stream = 'test', simulator.continuous = Simulator.Simple$new(), sleep = 1, iterations = Inf)}}{ 
    This method is used to create object of this class. It expects a host for
    a kafka cluster as input in the \code{kafka.host}, to which it can push
    the data.  Secondly, it expects the stream on the kafka host to listen on
    (\code{kafka.stream}), then it needs a simulator from which it can
    retrieve messages to push to the stream, and lastly, between each
    observation it will sleep for \code{sleep} number of seconds. Note that
    this function is currently not implemented due to the fact that kafka
    streaming has been removed for compatibility reasons.
 }
}
}

\keyword{datasets}
