% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSplitter.R
\docType{class}
\name{DataSplitter}
\alias{DataSplitter}
\title{DataSplitter}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DataSplitter
}
\description{
Splits data into train and test set.
}
\section{Methods}{
 
\describe{  
  \item{\code{initialize() }}{ 
    <<description>> 
  } 

  \item{\code{split(data) }}{ 
    splits the data in a train and test set
    @param data the data to spit 
  } 
}
}

\keyword{datasets}
