% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Density.Estimation.R
\docType{class}
\name{Density.Estimation}
\alias{Density.Estimation}
\title{Density.Estimation}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Density.Estimation
}
\description{
Density.Estimation
}
\keyword{datasets}
