% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DensityEstimation.R
\name{DensityEstimation.are_all_estimators_online}
\alias{DensityEstimation.are_all_estimators_online}
\title{Static method to check whether all included estimators are in fact specified
as being online estimators. If this is not the case, we should keep a cache
of all data somewhere.}
\usage{
DensityEstimation.are_all_estimators_online(estimators)
}
\arguments{
\item{estimators}{list of all estimator objects (should be density estimator objects)}
}
\value{
boolean TRUE if all estimators are online, FALSE if not
}
\description{
Static method to check whether all included estimators are in fact specified
as being online estimators. If this is not the case, we should keep a cache
of all data somewhere.
}
