% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DensityEstimation.R
\docType{class}
\name{DensityEstimation}
\alias{DensityEstimation}
\title{Density.Estimation
This class performs the actual density estimation for each of the random variables provided to it.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DensityEstimation
}
\description{
Density.Estimation
This class performs the actual density estimation for each of the random variables provided to it.
}
\keyword{datasets}
