% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.R
\name{Evaluation.accuracy}
\alias{Evaluation.accuracy}
\title{Evaluation.Accuracy}
\usage{
Evaluation.accuracy(data.observed, data.predicted)
}
\arguments{
\item{data.observed}{the true data (Y).}

\item{data.predicted}{the Y outcome from the estimator .}
}
\value{
double the calculated accuracy.
}
\description{
Function to calculate the accuract of an estimator, given its estimates and
the observed outcomes.
}
