% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.R
\name{Evaluation.get_evaluation_function}
\alias{Evaluation.get_evaluation_function}
\title{Returns the correct evaluation function given the family of the data
# @import Metrics}
\usage{
Evaluation.get_evaluation_function(family, useAsLoss = TRUE)
}
\arguments{
\item{family}{the family of the data (binomial for binary, gaussian for cts)}

\item{useAsLoss}{should we use the loss function or the performance function?}
}
\description{
Returns the correct evaluation function given the family of the data
# @import Metrics
}
