% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.R
\name{Evaluation.root_mean_squared_error}
\alias{Evaluation.root_mean_squared_error}
\title{RMSE caluclator}
\usage{
Evaluation.root_mean_squared_error(data.observed, data.predicted)
}
\arguments{
\item{data.observed}{the true data (Y)}

\item{data.predicted}{the Y outcome from the estimator}
}
\description{
RMSE caluclator
}
