% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_factory.R
\docType{methods}
\name{fabricate}
\alias{fabricate}
\alias{fabricate,LibraryFactory-method}
\title{Fabricate function}
\usage{
fabricate(object, SL.library)

\S4method{fabricate}{LibraryFactory}(object, SL.library)
}
\arguments{
\item{object}{An object}

\item{SL.library}{the library of different ML methods to use in the SL}
}
\description{
Fabricate function
}

