% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OnlineSuperLearner.S3.R
\name{fit.OnlineSuperLearner}
\alias{fit.OnlineSuperLearner}
\title{fit.OnlineSuperLearner}
\usage{
\method{fit}{OnlineSuperLearner}(formulae, data, algorithms = NULL,
  bounds = FALSE, ...)
}
\arguments{
\item{formulae}{list a list of all relevantVariable objects that need to be fitted}

\item{data}{data.frame or list of data.frames the data set to use for fitting the OSL}

\item{algorithms}{list of algorithms to use in the online superlearner}

\item{bounds}{either a list of bounds, or a boolean (default = FALSE), in
which TRUE forces the bounds to be generated automatically, FALSE causes the
bounds not to be generated at all (no normalization) we provide the option
to normalize the data in the OSL procedure. This entails that the package
will automatically select a set of bounds (min and max) based on the data
set provided. After that it will only use the normalized features (all
scaled between 0-1). The bounds should be specified as a list in which each
element represents one of the \code{RelevantVariable} objects. Each of these
entries should contain another list with two elements: \code{min_bound} and
\code{max_bound}, which represent the lower and upper bound of that
variable in specific.}

\item{...}{other parameters directly passed to the OSL and fit function.
There are several named variables to provide here:
- initial_data_size 
- max_iterations
- mini_batch_size
See for a full list the documentation of the \code{OnlineSuperLearner}
\code{fit} and \code{initialize} functions.}
}
\value{
a fitted version of an \code{OnlineSuperLearner} class
}
\description{
Fits an online superlearner using a similar notation as a GLM.
}
