% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global.R
\docType{data}
\name{generalImports}
\alias{generalImports}
\title{General packages used by all of the other classes}
\format{An object of class \code{list} of length 0.}
\usage{
generalImports
}
\description{
General packages used by all of the other classes
}
\keyword{datasets}
