% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H2O.initializer.R
\name{H2O.Available}
\alias{H2O.Available}
\title{Function to check whether the h2o cluster is already running}
\usage{
H2O.Available()
}
\value{
boolean TRUE if the cluster is running and FALSE if not.
}
\description{
Function to check whether the h2o cluster is already running
}
