% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterventionParser.R
\name{InterventionParser.first_intervention}
\alias{InterventionParser.first_intervention}
\title{InterventionParser.first_intervention}
\usage{
InterventionParser.first_intervention(intervention)
}
\arguments{
\item{intervention}{the intervention to give \code{list(variable, when, what)}.}
}
\value{
the first ever time at which an intervention is given.
}
\description{
Returns the time at which the first ever intervention should be given.
}
