% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterventionParser.R
\name{InterventionParser.parse_intervention}
\alias{InterventionParser.parse_intervention}
\title{InterventionParser.parse_intervention
Function to select the current intervention from an intervention list. It
will check whether the current timestamp is an intervention timestamp with
respect to the current outcome. If it is, it will return the actual
intervention.}
\usage{
InterventionParser.parse_intervention(intervention, current_time,
  current_outcome)
}
\arguments{
\item{intervention}{the intervention to give \code{list(variable, when, what)}}

\item{current_time}{integer with the current time}

\item{current_outcome}{the current random variable}
}
\value{
a list with when, what and if an intervention should be given at this time (\code{list(when, what, should_intervene)})
}
\description{
InterventionParser.parse_intervention
Function to select the current intervention from an intervention list. It
will check whether the current timestamp is an intervention timestamp with
respect to the current outcome. If it is, it will return the actual
intervention.
}
