% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterventionParser.R
\name{InterventionParser.valid_intervention}
\alias{InterventionParser.valid_intervention}
\title{InterventionParser.valid_intervention}
\usage{
InterventionParser.valid_intervention(intervention)
}
\arguments{
\item{intervention}{the intervention to check the validity of}
}
\value{
boolean \code{TRUE} if valid, \code{FALSE} if not
}
\description{
Checks whether the provided intervention is valid.
}
