% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{is.a}
\alias{is.a}
\title{Checks whether an object is an instance of the provided class}
\usage{
is.a(obj, obj.class)
}
\arguments{
\item{obj}{the object to check the class from}

\item{obj.class}{the class which we expect it to enherit from}
}
\description{
Checks whether an object is an instance of the provided class
}
