% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_factory.R
\docType{class}
\name{LibraryFactory-class}
\alias{LibraryFactory}
\alias{LibraryFactory-class}
\title{LibraryFactory}
\description{
This is the main super learner function
}
\section{Slots}{

\describe{
\item{\code{SL.type}}{the type of platform to use for fitting the models}
}}

