% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibraryFactory.R
\docType{class}
\name{LibraryFactory}
\alias{LibraryFactory}
\title{The libraryFactory can instantiate a set of models given to it.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LibraryFactory
}
\description{
The libraryFactory can instantiate a set of models given to it.
}
\section{Methods}{

\describe{
  \item{\code{new(ML.models.allowed = c('ML.H2O.glm', 'ML.Local.lm', 'ML.XGBoost.glm'))}}{
    This method is used to create object of this class. It expects a \code{ML.models.allowed} as a list which describes all of the models that are allowed for fabrication. 
    @param ML.models.allowed = the list of ML models allowed to create objects for. The default is a list of all models in the OSL package
    @return a new instance of the library factory
  }

  \item{\code{get_validity}}{
    Method to determine if the object is in a valid state. The method will throw whenever the state is not valid
  }

  \item{\code{fabricate(SL.library)}}{
    Method that fabricates the models in the provided \code{SL.library}.
    @param SL.library can be either a list of ML models, for which we will then choose the default hyper parameters,
    or a more specific specification of the models. This is in the form of a list and should be specified as follows:
    \code{list(list(algorithm = 'the class of the algorithm',
                    algorithm_params = list(hyper_parameter1=c(1,2,3)),
                    params = list(nbins = c(39, 40), online = FALSE))))} in which \code{algorithm_params} are the
                    hyperparameters for the learner, and params are the hyper parameters for the density estimator.
  }
}
}

\keyword{datasets}
