% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML.GLMnet.R
\docType{class}
\name{ML.GLMnet}
\alias{ML.GLMnet}
\title{ML.GLMnet}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ML.GLMnet
}
\description{
Wrapper for a glmnet estimator. From their discription:
Extremely efficient procedures for fitting the entire lasso or elastic-net
regularization path for linear regression, logistic and multinomial
regression models, Poisson regression and the Cox model. Two recent
additions are the multiple-response Gaussian, and the grouped multinomial
regression. The algorithm uses cyclical coordinate descent in a path-wise
fashion, as described in \url{http://www.jstatsoft.org/v33/i01}.
}
\section{Methods}{
 
\describe{  
  \item{\code{initialize(family = "binomial", alpha = 0.5, nlambda = 100, verbose = FALSE) }}{ 
    Creates a new GLMnet model

    @param family string (default = "binomial") the family to use for the
     estimator. Generally this should be binomial (since we are running
     various logistic regressions). However, if you have a good reason to
     use a different specification, change it here.

    @param alpha double (default = 0.5) the elasticnet mixing parameter, with $0 \le \alpha
     \le 1$. $\alpha=1$ is the lasso penalty, and $\alpha=0$ the ridge
     penalty.

    @param nlambda integer (default = 100) the number of lambda values.

    @param verbose (default = FALSE) the verbosity of the fitting procedure
  } 

  \item{\code{get_validity }}{ 
    Active method. Checks whether the specified configuration is valid.

    @return boolean, true if it is a valid configuration.
  } 
}
}

\keyword{datasets}
