% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML.H2O.glm.R
\docType{data}
\name{ML.H2O.glm}
\alias{ML.H2O.glm}
\title{ML.H2O.glm}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ML.H2O.glm
}
\description{
Wrapper for an H2O glm estimator. From their discription:
Generalized Linear Models (GLM) estimate regression models for outcomes
following exponential distributions. In addition to the Gaussian (i.e.
normal) distribution, these include Poisson, binomial, and gamma
distributions. Each serves a different purpose, and depending on
distribution and link function choice, can be used either for prediction or
classification.
}
\section{Methods}{
 
\describe{  
  \item{\code{initialize(nfolds = 1, alpha = 0.5, family = "gaussian") }}{ 
    Creates a new h2o glm model

    @param nfold integer (default = 1) specify the number of folds for
     cross-validation.
    @param alpha double (default = 0.5) the alpha parameter controls the distribution
     between the l1 (LASSO) and l2 (ridge regression) penalties. A value of
     1.0 for alpha represents LASSO, and an alpha value of 0.0 produces
     ridge reguression.

    @param family string (default = "binomial") the family to use for the
     estimator. Generally this should be binomial (since we are running
     various logistic regressions). However, if you have a good reason to
     use a different specification, change it here.
  } 
}
}

\keyword{datasets}
