% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML.H2O.randomForest.R
\docType{class}
\name{ML.H2O.randomForest}
\alias{ML.H2O.randomForest}
\title{ML.H2O.randomForest}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ML.H2O.randomForest
}
\description{
Wrapper for an H2O randomForest estimator. From their discription:
Distributed Random Forest (DRF) is a powerful classification and regression
tool. When given a set of data, DRF generates a forest of classification (or
regression) trees, rather than a single classification (or regression) tree.
Each of these trees is a weak learner built on a subset of rows and columns.
More trees will reduce the variance. Both classification and regression take
the average prediction over all of their trees to make a final prediction,
whether predicting for a class or numeric value (note: for a categorical
response column, DRF maps factors (e.g. ‘dog’, ‘cat’, ‘mouse) in
lexicographic order to a name lookup array with integer indices (e.g. ‘cat
-> 0, ‘dog’ -> 1, ‘mouse’ -> 2).
}
\section{Methods}{
 
\describe{  
  \item{\code{initialize(ntrees=50, min_rows=9) }}{ 
    Creates a new randomforest / drf model

    @param nfolds integer (default = 0) specify the number of folds for
     cross-validation.

    @param ntrees integer (default = 50) specify the number of trees to build.

    @param min_rows (default = 9) specify the minimum number of observations for a leaf (nodesize in R).

    @param verbose (default = FALSE) the verbosity of the fitting procedure
  } 
}
}

\keyword{datasets}
