% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML.H2O.R
\docType{class}
\name{ML.H2O}
\alias{ML.H2O}
\title{ML.H2O}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ML.H2O
}
\description{
Base class for any h2o machine learning model. Extend this class if you want
to create a new machine learning model.
}
\section{Methods}{
 
\describe{  
  \item{\code{initialize() }}{ 
    Creates a new ML.H2o base model. Note that this also initializes a new
    H2O interactor for storing and retrieving data.
  } 

  \item{\code{get_checkpoint(m.fit) }}{ 
    In order to perform an update on an existing H2O model, H2O uses
    checkpoints. In order to use this system together with our other
    implementations, we have this function to retrieve a checkpoint if one
    is available. The function returns \code{NULL} if no checkpoint is
    available, and it returns the model ID if there is a checkpoint.

    @param m.fit h2o model (not an h2o base model class, but a plain h2o
     model) a fitted h2o model from which one wants to retrieve a checkpoint.
  } 
}
}

\keyword{datasets}
