% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML.Local.Speedlm.R
\docType{class}
\name{ML.Local.Speedlm}
\alias{ML.Local.Speedlm}
\title{ML.Local.Speedlm}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ML.Local.Speedlm
}
\description{
Class to create speedlm linear models. Uses the speed glm package and can be
updated online in the case of the gaussian family. Note that the actual
online function is not yet implemented correctly. This requires us to use
the speedlm function, instead of the speed glm, wich does not support a
family for now.
}
\section{Methods}{

\describe{
  \item{\code{initialize()}}{
    Initializes a new speedlm estimator. 
  }
}
}

\keyword{datasets}
