% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML.randomForest.R
\docType{class}
\name{ML.randomForest}
\alias{ML.randomForest}
\title{ML.randomForest}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ML.randomForest
}
\description{
Class to create randomForest models
}
\section{Methods}{

\describe{
  \item{\code{initialize(ntrees = 500)}}{
    Initializes a new randomforest estimator. 
    @param ntrees integer (default = 500) the number of trees to use in the
     randomforest.
  }
}
}

\keyword{datasets}
