% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML.SpeedGLMSGD.R
\docType{class}
\name{ML.SpeedGLMSGD}
\alias{ML.SpeedGLMSGD}
\title{ML.SpeedGLMSGD}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ML.SpeedGLMSGD
}
\description{
Base class for Speedglm in combination with SGD machine learning model.
}
\keyword{datasets}
