% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML.SVM.R
\docType{class}
\name{ML.SVM}
\alias{ML.SVM}
\title{ML.SVM
Class to create support vector machine models}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ML.SVM
}
\description{
ML.SVM
Class to create support vector machine models
}
\section{Methods}{

\describe{
  \item{\code{initialize()}}{
    Initializes a new SVM estimator. 
  }
}
}

\keyword{datasets}
