% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML.XGBoost.glm.R, R/ML.XGBoost.gbm.R
\docType{class}
\name{ML.XGBoost.glm}
\alias{ML.XGBoost.glm}
\title{General Linear Model machine learning model, based on the XGBoost library}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ML.XGBoost.glm

ML.XGBoost.glm
}
\description{
General Linear Model machine learning model, based on the XGBoost library

General Linear Model machine learning model, based on the XGBoost library
}
\section{Fields}{

\describe{
\item{\code{model}}{the most recent / best model fitted.}

\item{\code{model}}{the most recent / best model fitted.}
}}
\section{Methods}{

\describe{
  \item{\code{new()}}{This method is used to create object of this class. }

  \item{\code{fit(data, X, Y)}}{Method to fit the linear model.}
}


\describe{
  \item{\code{new()}}{This method is used to create object of this class. }

  \item{\code{fit(data, X, Y)}}{Method to fit the linear model.}
}
}
\keyword{datasets}

