% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML.XGBoost.R
\docType{class}
\name{ML.XGBoost}
\alias{ML.XGBoost}
\title{Base class for any XGBoost machine learning model.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ML.XGBoost
}
\description{
Base class for any XGBoost machine learning model.
}
\keyword{datasets}

