% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MockFunction.R
\name{mock_function}
\alias{mock_function}
\title{A Mock function}
\usage{
mock_function(test_param = TRUE)
}
\arguments{
\item{test_param}{the parameter to test}
}
\description{
Function to test wether the package has been set up correctly
}
