% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_super_learner.R
\docType{class}
\name{OnlineSuperLearner-class}
\alias{OnlineSuperLearner}
\alias{OnlineSuperLearner-class}
\title{OnlineSuperLearner}
\value{
The best model
}
\description{
This is the main super learner function
}
\section{Slots}{

\describe{
\item{\code{X}}{A number}

\item{\code{Y}}{A number}
}}

