% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OnlineSuperLearner.R
\docType{data}
\name{OnlineSuperLearner}
\alias{OnlineSuperLearner}
\title{OnlineSuperLearner}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
OnlineSuperLearner
}
\description{
This is the main super learner class. This class contains everything related
to the super learner machine learning model.
}
\section{Fields}{

\describe{
\item{\code{SL.Library}}{the library of machinelearning algorithms to use}
}}
\section{Details}{

\code{$new(SL.Library)} starts a new OnlineSuperLearner. The provided /code{SL.Library}
contains the machine learning models to use

\code{$run()} runs the actual OnlineSuperLearning calculation

\code{$getModel()} returns the final OnlineSuperLearner model
}
\keyword{datasets}

