% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutputPlotGenerator.R
\name{OutputPlotGenerator.create_training_curve}
\alias{OutputPlotGenerator.create_training_curve}
\title{OutputPlotGenerator.create_training_curve
Function to plot the training curve given the number of iterations used}
\usage{
OutputPlotGenerator.create_training_curve(historical_cvs, randomVariables,
  output = "historical_cvs", dir = "~/tmp/osl/")
}
\arguments{
\item{historical_cvs}{the historical CV risks. List of lists of datatables.
First list is for each iteration, second for each learner, and the datatable
has columns for each random variable.}

\item{randomVariables}{list of randomvariables to use in the output}

\item{output}{string the filename to write the pdf to (without .pdf)}

\item{dir}{string the directory to write the file to}
}
\description{
OutputPlotGenerator.create_training_curve
Function to plot the training curve given the number of iterations used
}
