% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutputPlotGenerator.R
\name{OutputPlotGenerator.export_key_value}
\alias{OutputPlotGenerator.export_key_value}
\title{OutputPlotGenerator.export_key_value}
\usage{
OutputPlotGenerator.export_key_value(key, value, output = "variables",
  dir = "tmp")
}
\arguments{
\item{key}{the key to store the value under. That is, the data is stored as
a key value pair. This is the key in that pair.}

\item{value}{the value to store. That is, the data is stored as
a key value pair. This is the value in that pair.}

\item{output}{string (default = 'variables.dat') the filename to store the
data in.}

\item{dir}{string (default = 'tmp') the dir to write the file to.}
}
\description{
Function to export a key value pair to a file
}
