% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OnlineSuperLearner.R
\name{predict.OnlineSuperLearner}
\alias{predict.OnlineSuperLearner}
\title{Predict.OnlineSuperLearner}
\usage{
\method{predict}{OnlineSuperLearner}(object, newdata, Y, ...)
}
\arguments{
\item{object}{OnlineSuperLearner trained instance of an online superlearner class.}

\item{newdata}{data.base instance of an data.base class with new data to
perform the prediction with.}

\item{Y}{string the name of the Y variable to use as output of the prediction}

\item{...}{other parameters directly passed to the predict function}

\item{sample}{boolean would we like to sample an output or predict a probability}
}
\value{
data.table the predicted outcomes / probabilities
}
\description{
S3 prediction function for the online superlearner package. Can be used to
perform a prediction on the trained online superlearner object.
}
