% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessor.R
\docType{class}
\name{PreProcessor}
\alias{PreProcessor}
\title{PreProcessor}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PreProcessor
}
\description{
PreProcessor
}
\keyword{datasets}
