% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomVariable.R
\name{RandomVariable.find_ordering}
\alias{RandomVariable.find_ordering}
\title{Algorithm to find a possible ordering of the functions.
The worst case run time of this algorithm is pretty bad, and can it
probably done more efficiently}
\usage{
RandomVariable.find_ordering(randomVariables, verbose = FALSE)
}
\arguments{
\item{randomVariables}{the random variables to sort}

\item{verbose}{(default false) whether or not to be verbose when sorting}
}
\description{
Algorithm to find a possible ordering of the functions.
The worst case run time of this algorithm is pretty bad, and can it
probably done more efficiently
}
