% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomVariable.R
\docType{class}
\name{RandomVariable}
\alias{RandomVariable}
\title{RandomVariable}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RandomVariable
}
\description{
RandomVariable
}
\keyword{datasets}
