% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelevantVariable.R
\name{RelevantVariable.find_ordering}
\alias{RelevantVariable.find_ordering}
\title{Algorithm to find a possible ordering of the functions.
The worst case run time of this algorithm is pretty bad, and can it
probably done more efficiently}
\usage{
RelevantVariable.find_ordering(relevantVariables, verbose = FALSE)
}
\arguments{
\item{relevantVariables}{the relevant variables to sort}

\item{verbose}{(default false) whether or not to be verbose when sorting}
}
\description{
Algorithm to find a possible ordering of the functions.
The worst case run time of this algorithm is pretty bad, and can it
probably done more efficiently
}
