% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_super_learner.R
\docType{methods}
\name{run}
\alias{run}
\alias{run,OnlineSuperLearner-method}
\title{The run function}
\usage{
run(obj)

\S4method{run}{OnlineSuperLearner}(obj)
}
\arguments{
\item{obj}{The superlearner object}
}
\description{
The run function
}

