% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulator.Example.R
\docType{class}
\name{Simulator.Example}
\alias{Simulator.Example}
\title{Simulator.Example
The running example from the paper simulation}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Simulator.Example
}
\description{
Simulator.Example
The running example from the paper simulation
}
\section{Methods}{

\describe{
  \item{\code{new()}}{
    Starts a new simulator.
  }
  \item{\code{simulateWAY(numberOfBlocks, qw, ga, Qy, intervention, verbose)}}{
    Runs the simulation using the parameters provided.
  }
  \item{\code{simulateWAYiidTrajectories(numberOfBlocks, numberOfTrajectories, qw, ga, Qy, intervention, verbose)}}{
    does the exact same as simulateWAY, however, this function also takes an numberOfTrajectories argument, in which one can specify how many concurrent time series should be generated.
  }
}
}

\keyword{datasets}
