% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulator.RunningExample.R
\docType{class}
\name{Simulator.RunningExample}
\alias{Simulator.RunningExample}
\title{Simulator.RunningExample}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Simulator.RunningExample
}
\description{
The running example from the paper simulation: if and with how much a binary
treatment (antidepressant usage) attributes to changes in  depression.  We
use  a  normalized  hypothetical  set  of  sum  scores  of  the  Inventory
of DepressiveSymptomatology  (IDS) questionnaire with  a  possible  range
of 0-100.  Anti-depressant  usage  is a  binary  variable  sampled  from  a
binomial  distribution.
Variables:
w1 gender 0/1
w2 quality of sleep 0/1
w3 level of activity (PAL value) between 1.4 and 2.4
A antidepressant = 1 no antdepressant = 0
Y IDS (Inventory of Depressive Symptomatology)
Block time block
Patient id patient identification
Y_Y influence of Y on Y in the next block
Y_A influence of Y on A in the next block
Y_w2 influence of Y on w2 in the next blok
}
\section{Methods}{

\describe{
  \item{\code{new()}}{
    Starts a new simulator.
  }
  \item{\code{simulateWAY(numberOfBlocks, qw, ga, Qy, intervention, verbose)}}{
    Runs the simulation using the parameters provided.
  }
}
}

\keyword{datasets}
