% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulator.Simple.R
\docType{class}
\name{Simulator.Simple}
\alias{Simulator.Simple}
\title{Very basic simulator to generate data from a simple binary model}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Simulator.Simple
}
\description{
Very basic simulator to generate data from a simple binary model
}
\section{Fields}{

\describe{
\item{\code{model}}{the most recent / best model fitted.}
}}

\section{Methods}{

\describe{
  \item{\code{new()}}{Creates a new simulator}

  \item{\code{fit(nr.of.observations)}}{Generate a number of observations from the model (multple observations can be retrieved at once by using the \code{nr.of.observations} param}
}
}

\keyword{datasets}
