% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulator.Slow.R
\docType{class}
\name{Simulator.Slow}
\alias{Simulator.Slow}
\title{Simulator.Slow}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Simulator.Slow
}
\description{
Simulator.Slow
}
\keyword{datasets}
