% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMG.Base.R
\docType{class}
\name{SMG.Base}
\alias{SMG.Base}
\title{SMG.Base}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SMG.Base
}
\description{
Base class for any summary measure generator instance. These are used by the
\code{SummaryMeasureGenerator} to create summary measures.
}
\section{Methods}{

\describe{
  \item{\code{initialize()}}{
  }

  \item{\code{process()}}{
  }

  \item{\code{update(data.current)}}{
  }
}
}

\keyword{datasets}
