% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMG.Mean.R
\docType{class}
\name{SMG.Mean}
\alias{SMG.Mean}
\title{SMG.Mean}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SMG.Mean
}
\description{
The OnlineSuperLearner uses various \code{SummaryMeasureGenerator} instances
to generate new variables that can be used to represent the relevant history
of a variable. The \code{SMG.Mean} class allows a user to include the
running mean of a variable. The colname for such a variable is generated
with a _mean appendix.
}
\section{Methods}{
 
\describe{  
  \item{\code{initialize(colnames.to.mean) }}{ 
    Initializes a new \code{SMG.Mean} object.

    @param colnames.to.mean vector a vector containg the names of the
     variables for which to create a running mean variable.
  } 

  \item{\code{process(data.current) }}{ 
    Adds the running mean columns for each of the specified colnames.  It
    goes through the data and selects the variables specified on
    initialization.

    @param data.current data.table the currently available data.

    @return data.table with the new summary measure columns.
  } 

  \item{\code{update(data.current) }}{ 
    For online learning we need to be able to create new data blocks on the
    fly (as not all data is available beforehand. This function updates a
    set of variables to contain the new version of the running mean.

    @param data.current data.table the current data / the last data used for
     training, from which a new summary measure needs to be generated.

    @return data.table a new block for the next training iteration.
  } 

  \item{\code{exposedVariables}}{ 
    Active method. Returns a list of variables returned by this SMG.

    @return vector a vector of strings conaining all specified (used)
     contemporaneous colnames.
  } 

  \item{\code{minimalObservations}}{ 
    Active method. The minimal number of measurements needed for this SMG to
    be able to generate a new block. This is 1.

    @return integer the minimal number of measurements needed.
  } 
}
}

\keyword{datasets}
