% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMG.Mock.R
\docType{class}
\name{SMG.Mock}
\alias{SMG.Mock}
\title{SMG.Mock
Mock summary generator. For testing purposes only.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SMG.Mock
}
\description{
SMG.Mock
Mock summary generator. For testing purposes only.
}
\keyword{datasets}
