% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMG.Transformation.R
\docType{class}
\name{SMG.Transformation}
\alias{SMG.Transformation}
\title{SMG.Transformation
Transforms the current value of a variable using the given function}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SMG.Transformation
}
\description{
SMG.Transformation
Transforms the current value of a variable using the given function
}
\keyword{datasets}
