% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMGFactory.R
\docType{class}
\name{SMGFactory}
\alias{SMGFactory}
\title{SMGFactory}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SMGFactory
}
\description{
Factory for creating a set of SummaryMeasureGenerators, based on the needed variables
}
\section{Methods}{

\describe{
  \item{\code{initialize() }}{
    Creates a new SMG factory
  }

  \item{\code{fabricate(randomVariables, verbose = FALSE) }}{
    Fabricates the actual SMG's. Given a list of \code{RandomVariable}s all covariates are selected and merged.
    These are used as a basis for selecting the needed variables in the dataframe.
    @param randomVariables = a list of \code{RandomVariable} objects, from which the X variables are selected
    @param ... = data to get passed to the SMG
  }
}
}

\keyword{datasets}
