% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OnlineSuperLearner.S3.R
\name{summary.OnlineSuperLearner}
\alias{summary.OnlineSuperLearner}
\title{summary.OnlineSuperLearner}
\usage{
summary.OnlineSuperLearner(object, ...)
}
\arguments{
\item{object}{onlinesuperlearner the trained OSL instance}
}
\description{
S3 method to provide a summary about the online superlearner object. Prints
a description about the current fit of the OSL.
}
