% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryMeasureGenerator.R
\name{SummaryMeasureGenerator.generate_bounds}
\alias{SummaryMeasureGenerator.generate_bounds}
\title{Static function}
\usage{
SummaryMeasureGenerator.generate_bounds(data, eps = 0)
}
\arguments{
\item{data}{the data for which to generate the bounds}

\item{eps}{an extra margin to add to the estimated bounds}
}
\description{
Static function
}
