% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryMeasureGenerator.R
\docType{class}
\name{SummaryMeasureGenerator}
\alias{SummaryMeasureGenerator}
\title{SummaryMeasureGenerator}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SummaryMeasureGenerator
}
\description{
SummaryMeasureGenerator
}
\keyword{datasets}
