% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WCC.CVXR.R
\docType{class}
\name{WCC.CVXR}
\alias{WCC.CVXR}
\title{WCC.CVXR}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
WCC.CVXR
}
\description{
WCC.CVXR
}
\section{Methods}{

\describe{
  \item{\code{new()}}{
  }
  \item{\code{getWeigths()}}{
  }
  \item{\code{compute(Z, Y, libraryNames, weights.initial)}}{
  }
}
}

\keyword{datasets}
