% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WCC.NLopt.R
\docType{class}
\name{WCC.NLopt}
\alias{WCC.NLopt}
\title{WCC.NLopt}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
WCC.NLopt
}
\description{
WCC.NLopt
}
\section{Methods}{

\describe{
  \item{\code{new()}}{
   Creates a new computer for determining the best weighted combination of the ML libraries. This
   class uses the NLopt backend for optimizing
  }
  \item{\code{compute(Z, Y, libraryNames, obsWeights)}}{
   Method to compute the best weighted combination of the underlying estimators
   \code{Z} matrix containing the outcomes of each of the estimators
   \code{Y} vector containing the actual outcome
   \code{libraryNames} vector containing the names of the estimators
  }
}
}
\keyword{datasets}

