% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WCC.NNLS.R
\docType{class}
\name{WCC.NNLS}
\alias{WCC.NNLS}
\title{WCC.NNLS}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
WCC.NNLS
}
\description{
WCC.NNLS
}
\section{Methods}{

\describe{
  \item{\code{new()}}{
   Creates a new computer for determining the best weighted combination of the ML libraries.
   currently this is merely the NNLS version from the general superlearner, but in the future
   it should be updated to be 1. online, and 2. using better algorithm (like NLopts)
   \code{weights.initial} vector containing the initial weights
  }
  \item{\code{getWeigths()}}{
   Returns the current list of optimal weights (or the initial weights, if not yet fitted)
  }
  \item{\code{compute(Z, Y, libraryNames, weights.initial)}}{
   Method to compute the best weighted combination of the underlying estimators
   \code{Z} matrix containing the outcomes of each of the estimators
   \code{Y} vector containing the actual outcome
   \code{libraryNames} vector containing the names of the estimators
  }
}
}

\keyword{datasets}
