% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WCC.SGD.R
\docType{class}
\name{WCC.SGD}
\alias{WCC.SGD}
\title{WCC.SGD
Stochastic gradient descent optimizer, based on the R sgd package}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
WCC.SGD
}
\description{
WCC.SGD
Stochastic gradient descent optimizer, based on the R sgd package
}
\keyword{datasets}
