% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WCC.SGD.Simplex.R
\docType{class}
\name{WCC.SGD.Simplex}
\alias{WCC.SGD.Simplex}
\title{WCC.SGD.Simplex
This is the SGD computer used in the Online SL package by David Benkeser; https://github.com/benkeser/onlinesl/}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
WCC.SGD.Simplex
}
\description{
WCC.SGD.Simplex
This is the SGD computer used in the Online SL package by David Benkeser; https://github.com/benkeser/onlinesl/
}
\keyword{datasets}
