% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConstrainedGlm.R
\name{ConstrainedGlm.fit}
\alias{ConstrainedGlm.fit}
\title{Constrained logistic regression
In this function we create a regression for which the predicted probabilities are contstrained. That is, they can not be
less than a minimum of delta, or a maxiumum of 1 - delta.}
\usage{
ConstrainedGlm.fit(formula, delta, data, fall_back_to_glm = TRUE,
  previous_glm = NULL, ...)
}
\arguments{
\item{formula}{the formula to use for the regression}

\item{delta}{the threshold used for constraining the probabilities}

\item{...}{the other arguments passed to GLM}
}
\value{
a fitted glm
}
\description{
Constrained logistic regression
In this function we create a regression for which the predicted probabilities are contstrained. That is, they can not be
less than a minimum of delta, or a maxiumum of 1 - delta.
}
