% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossValidationRiskCalculator.R
\docType{class}
\name{CrossValidationRiskCalculator}
\alias{CrossValidationRiskCalculator}
\title{CrossValidationRiskCalculator}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CrossValidationRiskCalculator
}
\description{
CrossValidationRiskCalculator
}
\section{Methods}{

\describe{
  \item{\code{initialize() }}{
    <<description>>
  }

  \item{\code{calculate_risk(predicted.outcome, observed.outcome, randomVariables}}{
    <<description>>
    @param predicted.outcome
    @param observed.outcome
    @param randomVariables
  }

  \item{\code{update_risk(predicted.outcome = predicted.outcome, observed.outcome = observed.outcome}}{
    <<description>>
    @param predicted.outcome
    @param observed.outcome
  }

}
}

\keyword{datasets}
