% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.Base.R
\docType{class}
\name{Data.Base}
\alias{Data.Base}
\title{Base class for any data providing service class. Extend this class if you want to create a new method
to read data from}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Data.Base
}
\description{
Base class for any data providing service class. Extend this class if you want to create a new method
to read data from
}
\section{Methods}{

\describe{
  \item{\code{initialize() }}{
    Creates a new instance of the data.base class
  }

  \item{\code{getNext()}}{
    This method returns the next observation from the underlying data, depending on the implementation this could be
    a row from a data frame, or an observation from a stream. Note: this method needs to be inherited in the
    subclass.
  }

  \item{\code{getNextN(n = 1)}}{
    This method returns the next \code{n} observations from the underlying data, depending on the implementation this
    could be a number of rows from a data frame, or a number of observations from a stream. Note: this method needs
    to be inherited in the subclass.
    @param n the number of elements to retrieve.
  }
}
}

\keyword{datasets}
