% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.Static.R
\docType{class}
\name{Data.Static}
\alias{Data.Static}
\title{Data.Static}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Data.Static
}
\description{
Data.Static
}
\keyword{datasets}
