% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.Static.R
\docType{class}
\name{Data.Static}
\alias{Data.Static}
\title{Class to provide the functionality of treating a static, finite datatable as a stream of incomming
data.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Data.Static
}
\description{
Class to provide the functionality of treating a static, finite datatable as a stream of incomming
data.
}
\section{Methods}{

\describe{
  \item{\code{new(dataset, url)}}{
    This method is used to create object of this class. One can provide either a \code{datatable} or a \code{url} pointing to a dataframe.
  }

  \item{\code{getAll()}}{
    Method to retrieve the whole dataset at once.
  }

  \item{\code{getNext()}}{
    Method to retrieve the next observation from the data.
  }

  \item{\code{getNextN(n)}}{
    Method that returns the next \code{n} observations. This function can be used to bootstrap an initial model.
  }
}
}

\keyword{datasets}
