% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.Stream.R
\docType{class}
\name{Data.Stream}
\alias{Data.Stream}
\title{Data.Stream}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Data.Stream
}
\description{
Class to provide the functionality to read data from an input stream.
}
\section{Methods}{

\describe{
  \item{\code{initialize(kafka.host = 'imac.evionix.org:2181', kafka.stream = 'test')}}{
   This method is used to create object of this class. It expects a host for
   a kafka cluster (or the coordinating zookeeper therof) as input in the
   \code{kafka.host}. Secondly, it also expects the stream on the kafka host
   to listen on (\code{kafka.stream}. Warning! This function is currently
   not implemented due to difficulties with the kafka R package (it needs
   java and rjava, which causes problems on some systmes). Although
   implementation is not a difficult task, it is currently removed for
   compatibility reasons. One can recreate this function / class when
   needed.
   @param kafka.host the host address for the kafka stream
   @param kafka.stream the particular stream to read from
  }
  \item{\code{getNext()}}{
    Currently not implemented (see \code{initialize} for reasons)
    Returns the next observation from the stream
    @return one row of the dataframe the stream yields
  }

  \item{\code{getNextN(n=1)}}{
    Currently not implemented (see \code{initialize} for reasons)
    Returns the next \code{n} observations from the stream
    @return data.table a number of \code{n} rows concatenated in a data.table
  }

  \item{\code{getNext()}}{Method to retrieve the next observation from the stream.}
}
}

\keyword{datasets}
