% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.Stream.R
\docType{data}
\name{Data.Stream}
\alias{Data.Stream}
\title{Data.Stream}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Data.Stream
}
\description{
Data.Stream
}
\keyword{datasets}

