% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataCache.R
\docType{class}
\name{DataCache}
\alias{DataCache}
\title{DataCache
Class that is used internally to manage datacaching}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DataCache
}
\description{
DataCache
Class that is used internally to manage datacaching
}
\keyword{datasets}
