% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSplitter.R
\docType{class}
\name{DataSplitter}
\alias{DataSplitter}
\title{DataSplitter}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DataSplitter
}
\description{
Splits data into train and test set.
}
\keyword{datasets}
