% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSplitter.R
\docType{class}
\name{DataSplitter}
\alias{DataSplitter}
\title{DataSplitter}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DataSplitter
}
\description{
Splits a dataset into a train and test set.
}
\section{Methods}{
 
\describe{  
  \item{\code{initialize(test_set_size = 1) }}{ 
    Creates a new datasplitter

    @param test_set_size (default = 1) an integer to specify the size of the
     test set to use.
  } 

  \item{\code{split(data)}}{ 
    Splits the data in a train and test set. A block that was used as a test
    set previously is automatically appended to the traininset (so all data
    gets used for training that has been used for testing before).

    In the current implementation it will always use the last observation as
    the test set observation.  Eventually this should be configurable, and
    \code{n} number of observations should be includable.

    @param data data.table the data to spit into a train and test set.

    @param test_set_size integer (default = NULL) overrides the initialized
     test_set_size. If \code{NULL}, we use the initialized one (default).

    @return list with two entries: \code{train} and \code{test}. Each
     containing the respective dataframe.
  } 

  \item{\code{get_test_set_size}}{ 
    Active method. The size of the testset used by this splitter instance.
  } 
}
}

\keyword{datasets}
