% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.R
\name{Evaluation.accuracy}
\alias{Evaluation.accuracy}
\title{Accuracy caluclator}
\usage{
Evaluation.accuracy(data.observed, data.predicted)
}
\arguments{
\item{data.observed}{the true data (Y)}

\item{data.predicted}{the Y outcome from the estimator}
}
\description{
Accuracy caluclator
}
