% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.Accuracy.R
\docType{class}
\name{Evaluation.Accuracy}
\alias{Evaluation.Accuracy}
\title{Evaluation.Accuracy caluclator}
\usage{
Evaluation.Accuracy(model, data, Y, A, W)
}
\description{
Evaluation.Accuracy caluclator
}

