% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.R
\name{Evaluation.log_likelihood_loss}
\alias{Evaluation.log_likelihood_loss}
\title{Log likelihood loss evaluation metric}
\usage{
Evaluation.log_likelihood_loss(data.observed = NULL, data.predicted,
  eps = 1e-15)
}
\arguments{
\item{data.observed}{is unused. It is included in the function to comply to the general}

\item{data.predicted}{the Y outcome from the estimator}

\item{eps}{is a small offset to let the log not go to Inf}
}
\description{
Log likelihood loss evaluation metric
}
