% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.R
\name{Evaluation.mse_loss}
\alias{Evaluation.mse_loss}
\title{Evaluation.mse_loss}
\usage{
Evaluation.mse_loss(data.observed, data.predicted)
}
\arguments{
\item{data.observed}{the true data (Y)}

\item{data.predicted}{the Y outcome from the estimator}
}
\value{
double the calculated mean squared error loss
}
\description{
Function to calculate the mean squared error loss
}
