% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.R
\name{Evaluation.root_mean_squared_error}
\alias{Evaluation.root_mean_squared_error}
\title{RMSE caluclator}
\usage{
Evaluation.root_mean_squared_error(data.observed, data.predicted)
}
\description{
RMSE caluclator
}
