% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.R
\name{Evaluation.root_mean_squared_error}
\alias{Evaluation.root_mean_squared_error}
\title{Evaluation.root_mean_squared_error}
\usage{
Evaluation.root_mean_squared_error(data.observed, data.predicted)
}
\arguments{
\item{data.observed}{the true data (Y)}

\item{data.predicted}{the Y outcome from the estimator}
}
\value{
double the root of the mean squared error.
}
\description{
Function to calculate the root of the mean squared error.
}
