% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OnlineSuperLearner.S3.R
\name{fit.OnlineSuperLearner}
\alias{fit.OnlineSuperLearner}
\title{fit.OnlineSuperLearner}
\usage{
fit.OnlineSuperLearner(formulae, data, algorithms = NULL, normalize = FALSE,
  measurements_per_obs = Inf, ...)
}
\arguments{
\item{formulae}{list a list of all randomVariable objects that need to be fitted}

\item{data}{data.frame the data set to use for fitting the OSL}

\item{algorithms}{list of algorithms to use in the online superlearner}

\item{normalize}{boolean (default = FALSE) we provide the option to
normalize the data in the OSL procedure. This entails that the package will
automatically select a set of bounds (min and max) based on the data set
provided. After that it will only use the normalized features (all scaled
between 0-1).}

\item{measurements_per_obs}{integer (default = Inf) the number of
measurments in a single observation.}

\item{...}{other parameters directly passed to the OSL and fit function.
There are several named variables to provide here:
- initial_data_size 
- max_iterations
- mini_batch_size
See for a full list the documentation of the \code{OnlineSuperLearner}
\code{fit} and \code{initialize} functions.}
}
\value{
a fitted version of an \code{OnlineSuperLearner} class
}
\description{
Fits an online superlearner using a similar notation as a GLM.
}
