% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H2O.initializer.R
\docType{methods}
\name{H2O.initializer}
\alias{H2O.initializer}
\alias{H2O.initializer,ANY-method}
\title{H2O.initializer}
\usage{
H2O.initializer(host = "localhost", port = 54321, runlocal = TRUE)

H2O.initializer(host = "localhost", port = 54321, runlocal = TRUE)

\S4method{H2O.initializer}{ANY}(host = "localhost", port = 54321,
  runlocal = TRUE)
}
\arguments{
\item{host}{the host h2o is running on (defaults to localhost)}

\item{port}{the port h2o runs on (defaults to 54321)}

\item{runlocal}{whether or not we should start h2o, or whether it is running somewhere already}
}
\value{
an instance of H2O.initializer
}
\description{
H2O.initializer

H2O.initializer constructor function
}

