% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H2O.Initializer.R
\docType{class}
\name{H2O.Initializer}
\alias{H2O.Initializer}
\title{H2O.Initializer}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
H2O.Initializer
}
\description{
H2O.Initializer
}
\keyword{datasets}

