% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H2O.initializer.R
\docType{class}
\name{H2O.Initializer}
\alias{H2O.Initializer}
\title{H2O.Initializer
Idealy I'd like to use a singletonpattern for this, but this also works.}
\usage{
H2O.Initializer(host = "localhost", port = 54321, runlocal = TRUE,
  verbose = FALSE)
}
\description{
H2O.Initializer
Idealy I'd like to use a singletonpattern for this, but this also works.
}
