% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H2O.Interactor.R
\docType{class}
\name{H2O.Interactor}
\alias{H2O.Interactor}
\title{H2O.Interactor}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
H2O.Interactor
}
\description{
H2O.Interactor
}
\keyword{datasets}
