% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterventionParser.R
\name{InterventionParser.parse_intervention}
\alias{InterventionParser.parse_intervention}
\title{InterventionParser.parse_intervention}
\usage{
InterventionParser.parse_intervention(intervention, current_time,
  current_outcome, check = FALSE)
}
\arguments{
\item{intervention}{the intervention to give \code{list(variable, when, what)}}

\item{current_time}{integer with the current time}

\item{current_outcome}{string the name of the current relevant variable}

\item{check}{boolean perform checks on the input variables}
}
\value{
a list with when, what and if an intervention should be given at
 this time (\code{list(when, what, should_intervene)}). If one whould not
 intervene, the when and what are -1.
}
\description{
Function to select the current intervention from an intervention list. It
will check whether the current timestamp is an intervention timestamp with
respect to the current outcome. If it is, it will return the actual
intervention. The intervention needs to be specified as follows. Each
intervention is a list consisting of three attributes: \code{variable},
\code{when}, and \code{what}. The \code{variable} entry denotes the name of
the variable on which the intervention is done. Usually this is the 'A'
variable, but this could be tweaked. Secondly one has to specify the
\code{when} variable. This variable denotes when the intervention should be
performed. Lastly the \code{what} variable. This denotes the actual
intervention to be performed. E.g., one could denote treatment as 1 and
control as 0. Then an intervention could be to give everybody a treatment,
forcing everyone to have a 1 as treatment variable. This would then boil
down to a \code{what} of 1.
}
