% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterventionParser.R
\name{InterventionParser.valid_intervention}
\alias{InterventionParser.valid_intervention}
\title{InterventionParser.valid_intervention
Checks whether the provided intervention is valid}
\usage{
InterventionParser.valid_intervention(intervention)
}
\arguments{
\item{intervention}{the intervention to check the validity of}
}
\value{
boolean true if valid, false if not
}
\description{
InterventionParser.valid_intervention
Checks whether the provided intervention is valid
}
