% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibraryFactory.R
\docType{class}
\name{LibraryFactory}
\alias{LibraryFactory}
\title{LibraryFactory}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LibraryFactory
}
\description{
LibraryFactory
}
\section{Fields}{

\describe{
\item{\code{ML.Local}}{the type of platform to use for fitting the models}
}}

\keyword{datasets}
