% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML.Local.lm.R
\docType{class}
\name{ML.Local.lm}
\alias{ML.Local.lm}
\title{ML.Local.lm}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ML.Local.lm
}
\description{
ML.Local.lm
}
\keyword{datasets}
