% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML.Local.Speedlm.R
\docType{class}
\name{ML.Local.Speedlm}
\alias{ML.Local.Speedlm}
\title{ML.Local.Speedlm
Class to create speedlm linear models. Uses the speed glm package and can be updated online.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ML.Local.Speedlm
}
\description{
ML.Local.Speedlm
Class to create speedlm linear models. Uses the speed glm package and can be updated online.
}
\section{Methods}{

\describe{
  \item{\code{initialize()}}{
    Initializes a new speedlm estimator. 
  }
}
}

\keyword{datasets}
