% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML.NeuralNet.R
\docType{class}
\name{ML.NeuralNet}
\alias{ML.NeuralNet}
\title{ML.NeuralNet
Class to create neuralnetworks}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ML.NeuralNet
}
\description{
ML.NeuralNet
Class to create neuralnetworks
}
\section{Methods}{

\describe{
  \item{\code{initialize(hidden)}}{
    Initializes a new neuralnet estimator with the provided hidden layers. 
  }
}
}

\keyword{datasets}
