% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML.randomForest.R
\docType{class}
\name{ML.randomForest}
\alias{ML.randomForest}
\title{ML.randomForest
Class to create randomForest models}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ML.randomForest
}
\description{
ML.randomForest
Class to create randomForest models
}
\section{Methods}{

\describe{
  \item{\code{initialize()}}{
    Initializes a new randomforest estimator. 
  }
}
}

\keyword{datasets}
