% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneStepEstimator.R
\docType{class}
\name{OneStepEstimator}
\alias{OneStepEstimator}
\title{OneStepEstimator}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
OneStepEstimator
}
\description{
OneStepEstimator
}
\keyword{datasets}
