% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OnlineSuperLearner.R
\docType{class}
\name{OnlineSuperLearner}
\alias{OnlineSuperLearner}
\title{OnlineSuperLearner}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
OnlineSuperLearner
}
\description{
This is the main super learner class. This class contains everything related
to the super learner machine learning model.
}
\section{Fields}{

\describe{
\item{\code{SL.Library}}{the library of machinelearning algorithms to use}
}}

\section{Details}{

\code{$new(SL.Library)} starts a new OnlineSuperLearner. The provided /code{SL.Library}
contains the machine learning models to use
\code{$run()} runs the actual OnlineSuperLearning calculation
\code{$getModel()} returns the final OnlineSuperLearner model
\code{$getModel(data, X, Y, onlySL)} allows the user to make predictions on new data using
the fitted superlearner model. The data provided in this function is the new data object,
which can be either a stream of data or a dataframe. X is a vector of covariate names to use for the
prediction, Y is a vector containing the variables to predict
}

\keyword{datasets}
