% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OnlineSuperLearner.Simulation.R
\docType{class}
\name{OnlineSuperLearner.Simulation}
\alias{OnlineSuperLearner.Simulation}
\title{OnlineSuperLearner.Simulation runs different simulations on the superlearner object.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
OnlineSuperLearner.Simulation
}
\description{
OnlineSuperLearner.Simulation runs different simulations on the superlearner object.
}
\keyword{datasets}
