% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutputPlotGenerator.R
\name{OutputPlotGenerator.create_convergence_plot}
\alias{OutputPlotGenerator.create_convergence_plot}
\title{OutputPlotGenerator.create_convergence_plot
Function to plot the convergence to the parameter of interest of both the truth and the estimation.}
\usage{
OutputPlotGenerator.create_convergence_plot(data, output,
  dir = "~/tmp/osl/convergence")
}
\arguments{
\item{data}{the truth  and apporoximations to plot}

\item{output}{the filename of the output}

\item{dir}{the directory name of the output}
}
\description{
OutputPlotGenerator.create_convergence_plot
Function to plot the convergence to the parameter of interest of both the truth and the estimation.
}
