% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutputPlotGenerator.R
\name{OutputPlotGenerator.create_convergence_plot}
\alias{OutputPlotGenerator.create_convergence_plot}
\title{OutputPlotGenerator.create_convergence_plot
Function to plot the convergence to the parameter of interest of both the truth and the estimation.}
\usage{
OutputPlotGenerator.create_convergence_plot(data, output, dir = "~/tmp/osl/")
}
\arguments{
\item{truth_approximation}{the truth (i.e., as found using simulation / monte carlo approximation)}

\item{estimated_approximation}{the estimate (i.e., as found using machine learning and monte carlo approximation)}
}
\description{
OutputPlotGenerator.create_convergence_plot
Function to plot the convergence to the parameter of interest of both the truth and the estimation.
}
