% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutputPlotGenerator.R
\name{OutputPlotGenerator.create_risk_plot}
\alias{OutputPlotGenerator.create_risk_plot}
\title{OutputPlotGenerator.create_risk_plot
Function to create plots similar to the ones in the OSL papers}
\usage{
OutputPlotGenerator.create_risk_plot(performance, output, dir = "~/tmp/osl")
}
\arguments{
\item{performance}{the performances as as calculated using crossvalidation. See the \code{CrossValidationRiskCalculator} class.}

\item{output}{the name of the output file}

\item{dir}{the dir to write the output file to}
}
\description{
OutputPlotGenerator.create_risk_plot
Function to create plots similar to the ones in the OSL papers
}
