% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutputPlotGenerator.R
\name{OutputPlotGenerator.create_training_curve}
\alias{OutputPlotGenerator.create_training_curve}
\title{OutputPlotGenerator.create_training_curve}
\usage{
OutputPlotGenerator.create_training_curve(historical_cvs, relevantVariables,
  output = "historical_cvs", dir = "~/tmp/osl/")
}
\arguments{
\item{historical_cvs}{the historical CV risks. List of lists of datatables.
First list is for each iteration, second for each learner, and the datatable
has columns for each relevant variable.}

\item{relevantVariables}{list of relevantvariables to use in the output}

\item{output}{string (default = 'historical_cvs') the filename to write the
pdf to (without .pdf).}

\item{dir}{string (default = '~/tmp/osl/') the directory to write the file
to.}
}
\description{
Function to plot the training curve given the number of iterations used
}
