% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutputPlotGenerator.R
\name{OutputPlotGenerator.export_key_value}
\alias{OutputPlotGenerator.export_key_value}
\title{OutputPlotGenerator.export_key_value 
Function to export a key value pair to a file}
\usage{
OutputPlotGenerator.export_key_value(key, value, dir = "~/tmp/osl")
}
\arguments{
\item{key}{the key to store the value under}

\item{value}{the value to store}

\item{dir}{the dir to write the file to}
}
\description{
OutputPlotGenerator.export_key_value 
Function to export a key value pair to a file
}
