% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OnlineSuperLearner.S3.R
\name{predict.OnlineSuperLearner}
\alias{predict.OnlineSuperLearner}
\title{predict.OnlineSuperLearner}
\usage{
\method{predict}{OnlineSuperLearner}(object, newdata, Y = NULL, ...)
}
\arguments{
\item{object}{OnlineSuperLearner trained instance of an online superlearner class.}

\item{newdata}{the new data to perform the prediction with. Note that this
can be a data.frame, after which we will generate blocks based on the
measurements in the data, or a \code{Data.Base}, which _should_ already
include all necessary variables.}

\item{Y}{the dependent variables for which we want to predict the outcome.
The parameter is allowed to take several forms:
 - List of \code{RandomVariable} objects to predict
 - Single \code{RandomVariable} object to predict
 - List of strings with the names of the outputs (\code{list('X','Y')})
 - Single string with the name of the output (\code{'Y'})}

\item{...}{other parameters directly passed to the predict function}
}
\value{
\code{data.table} a \code{data.table} containing the predicted probabilities
}
\description{
S3 prediction function for the online superlearner package. Can be used to
perform a prediction on the trained online superlearner object.
}
