% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulator.GAD.R
\docType{class}
\name{Simulator.GAD}
\alias{Simulator.GAD}
\title{Simulator.GAD
directed acyclic graph approach to simulation}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Simulator.GAD
}
\description{
Simulator.GAD
directed acyclic graph approach to simulation
}
\keyword{datasets}
