% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMG.Latest.Entry.R
\docType{class}
\name{SMG.Latest.Entry}
\alias{SMG.Latest.Entry}
\title{SMG.Latest.Entry}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SMG.Latest.Entry
}
\description{
SMG.Latest.Entry
}
\keyword{datasets}
