% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMG.Transformation.R
\docType{class}
\name{SMG.Transformation}
\alias{SMG.Transformation}
\title{SMG.Transformation
Transforms the current value of a variable using the given function}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SMG.Transformation
}
\description{
SMG.Transformation
Transforms the current value of a variable using the given function
}
\section{Methods}{
 
\describe{  
  \item{\code{initialize(function_to_use = sin, suffix = "sin", colnames_to_use) }}{ 
    <<description>> 
    @param function_to_use = sin 
    @param suffix = "sin" 
    @param colnames_to_use  
  } 

  \item{\code{update(data.current) }}{ 
    <<description>> 
    @param data.current  
  } 

  \item{\code{process(data.current) }}{ 
    <<description>> 
    @param data.current  
  } 
}
}

\keyword{datasets}
