% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WCC.NLopt.R
\docType{class}
\name{WCC.NLopt}
\alias{WCC.NLopt}
\title{WCC.NLopt}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
WCC.NLopt
}
\description{
WCC.NLopt
}
\section{Methods}{

\describe{
  \item{\code{new()}}{
   Creates a new computer for determining the best weighted combination of the ML libraries. This
   class uses the NLopt backend for optimizing
   \code{lossFunction} function (optional) function to compute the loss of the OSL
   \code{lossFunctionGradient} function (optional) gradient of the function to compute the loss for the OSL
   \code{opts} list (optional) list with options for 
  }
  \item{\code{getWeigths()}}{
   Returns the current list of optimal weights (or the initial weights, if not yet fitted)
  }
  \item{\code{compute(Z, Y, libraryNames)}}{
   Method to compute the best weighted combination of the underlying estimators. The function will use the weights supplied to the constructor. If the model has been fitted before, it will reuse the weights of the previous run as the starting value for the current run.
   \code{Z} matrix containing the outcomes of each of the estimators
   \code{Y} vector containing the actual outcome
   \code{libraryNames} vector containing the names of the estimators
  }
}
}
\keyword{datasets}

